Refactored geometric initialization into a seperate base class "Triangular2DFEM" from which "Elasticity2DSolver" inherits. Also added functionality for zero-Dirichlet BC's when solving for eigenvibration modes/frequencies.
