+++ root/__main__.py
@@ -15,7 +15,7 @@ def main():
     # test_markov(N=10, area="plate")
     # test_animate_mesh_stress(N=10)
     # save_material_vibration_modes(N=6, k_min=3, k_max=4)
-    test_full_solver(N=10)
+    test_full_solver(N=12)
 
     pass
